namespace assets under 'romo' subfolder
This adds the namespace subfolder 'romo' to the asset paths.  This
is a best-practive measure when developing 3rd-party gems that will
serve assets with Dassets.  This keeps the global "css" or "js" root
namespace but adds in the gem specific "css/romo" or "js/romo"
namespaces.  This way different gems can provide assets with the
same name and they won't stomp each other.