base js with init ui callback behavior
This adds a base js file to do basic Romo js handling.  The first
basic behavior added here is the ability to add callbacks to init
UI.  These callbacks will be triggered any time new html is loaded.
Right now, this means on page load and when the `initHtml` helper
is used.

Plugins and components will use this system to auto initialize their
UI.  This keeps the user from having to remember to init UI everytime
a page is loaded or new UI is loaded after the page load.