tweaks to make data attribute detection more flexible
This updates the js components to make them use a more flexible
method for determining when attributes are missing.  In addtion to
not being defined, attributes will be treated as missing if their
values are empty.

This helps when you have component markup rendered from a partial.
The partial may populate these attrs from a variable and therefore
always define the attr and making it empty if the var is empty or
missing.