tooltip: rewrite to dynamically build the popup
This switches the tooltip implementation to dynamically build the
popup element and attach it to the body.  The benefit here is the
probability for unintentional styling goes way down and there is
less markup for the developer to add.

The downside of this approach is that we now have to absolutely
position the popup relative the the toggle element.  This involves
doing the math to center-position it based on the given position
attr and then binding to window resizes to adjust the position.
This also assumes the tooltip is show and has its final content
in place before positioning it.  This also means we have to re-
position the tooltip every time it is opened.

Using this approach eleminates most of the edge-case scenarios of
the previous approach and allows for a better developer experience.