dropdown: rework to dynamically add popup and position
This reworks the dropdown component to add its popup and position
it in the same fashion that the tooltip does.  This has the advantage
of the popups being attached to the body to minimize unintentional
styles and negating the need some relative positioning wrapper elem.

This also generally cleans up the js and css based on the new design
and some conventions set in the tooltip js.