invoke component
This adds a new component for invoking action on a click event on
a given element.  It also handles optionally loading data via a
given `href` attribute on the invoke element.

This is a common pattern in Romo components: have a link/button
that should, on click, make an ajax call to some href and do something
with the data.  This pattern is used on the Dropdown component and
will be used on the coming Inline and Modal components.  This may
be used eventually on tooltips that load dynamic content some day
as well.

This is prep for adding the Inline and Modal components, specifically.