dropdown: rework the body load triggers
This moves the body load event triggers out of event handlers and
into the corresponding action methods.  This ensures that the events
will always be triggered, whether via another event or manuall calls
to the action methods.

This also removed the comments recommending you override the action
methods.  A better approach is to bind to the events the action
methods trigger in a custom `doInit()` function.