form: basic first pass
This adds a component for doing ajax form submissions.  This handles
creating the form and hooking up the submission to an ajax request.
It also posts events to handle state changes and errors.

This is a first pass - it really only handles the ajax stuff and
submitting when hitting 'Enter' on anything but a text area.  In
future changes, I'll add more integrated error handling defaults,
list values input handling, and param encoding options.

I chose to do lose error handling to start as most error handling
is heavily bound to the UI you are putting in your form and I wanted
to keep this first pass pretty UI-agnostic.  Once Romo's UI handling
and CSS/structure mature's I plan on coming back and doing some
default error handling logic.

Note too, I switched the event name convention to _not_ prepend the
names with "on".  This is redundant when binding them:
`elem.on('channel:onEvent').  I'll roll out this convention change
to the other components in future changes.