form: reload page option
This adds a `data-form-reload-page` option to the form.  Set this
to true and the form will automatically reload the current page
you are on after the form successfully submits.

This helps with the common pattern of: submit a form to make some
changes and then reload the page to see those changes.  It makes
use of the base `Romo.reloadPage` helper to do the page reload.