form: process both ajax and redirect GET submits
This updates the form's submit behavior to no only process ajax
get requests but also more standard get submits that should redirect
the page.

Use the `data-form-redirect-page` option on a form where the method
is GET and that form will, instead of the normal ajax submission,
redirect the page to the action/params specified by the form.  Use
this where you want the form to redirect to an action - for example
in search forms.