form: optional list input values with a delimeter
This updates how Romo forms serialize their parameter values to
allow for opitonally listing values with a delimeter.  The best
way to illustrate this is with an example:

Say you have a multi-select (or some other multi-input like checkboxes)
named `type`.  Normally, when type's values are serialized they
would look something like `type[]=val1&type[]=val2`.  This change
allows you to tell the input, using a `data-form-list-values="true"`
attr, to list the values in a single string value using a delimeter.
So this would result in the input serializing as `type=val1,val2`.
By default `,` is used as the delimeter - override that using the
`data-form-list-values-delim` attr.

This behavior is of course optional - values serialize normally
if you don't use the list values attr.