switch to Romo namespaces data attributes
This switches to using 'romo-' prefixed data attributes.  This has
the effect of namespacing all romo data attributes.  This is in an
effort to be a good citizen and play nice with other UI kits.

The other big change here is how the auto-init system detects elems
to auto-init.  Before, it would require a class _plus_ a data attribute.
This seemed excessive and had the negative side-effect of requiring
you to use a css class when you may not wish to.  With namespaced
data attrs, relying soley on the data attr to detect is sufficient.