invoke: make the event to trigger the invoke variable
It still defaults to `'click'` events to trigger invokes, but you
can now specify a different trigger, like 'hover' or whatever, to
trigger the event.

This also breaks out do methods for binding and unbinding the invoke
event.

These features are all geared towards the coming inline component.
The design will be to pair an inline with an invoke and then use
the invoke to trigger inline behavior.