invoke: honor data loading when triggering 'invoke' event
This has the component be more smart about when it fires the 'invoke'
event.  If there is data to load, it will wait to fire the invoke
event until _after_ the data has loaded or has failed to load.  If
there is no data to load, the event is fired immediately.

This aids things (like inlines) that need to show/hide UI but need
the data fully loaded beforing showing/hiding the UI.