form: cleanup init params and name more explicitly
The form component allows you to pass an optional given submit
element.  It prefers any given submit element but will fall back
to its default submit element: a standard submit nested in the form.
This updates the var names to be more explicit about which is which
and how they interact.

This also recognizes any element with `data-romo-form-submit="true"`
as an attr as a default submit elem.  This adds fleibility when you
don't want to use a traditional submit input to submit your forms.