inlineForm: first pass
This component is a "mash-up" of an inline and a form.  It designed
to be used for inline editing - submitting forms from an inline
component.

The key here is that an inlineForm component is both an inline *and*
a form.  Events from both components are triggered on it.  This is
particularly advantageous because you can bind to form events on
a form that is dynamically loaded in the inline.

This component handles initializing the inline component automatically
for you.  If you want any form in the inline to be auto initialized, add
the `data-romo-form-auto="inlineForm"` attribute to them.  It attempts
to auto initialize its form on both init and on inline load success.