inlineForm: trigger inlineForm-specific proxy events
This switches to triggering form-proxy-events with an "inlineForm"
namespace.  Since the form is a child of the inlineForm, it's events
are already published on the parent inlineForm element.  Our proxy
events are special in that they trigger with the form in the args;
they need their own namespaced name to distinguish them from the
standard form events.