modal: first pass, modeled after dropdowns
This is the first pass at the modal component.  It is modeled after
dropdowns in terms of how it is invoked, its parts, and how it
loads body content.

A modal is really just a dropdown that isn't anchored to the trigger
elem, but instead is anchored (fixed) to the center of the viewport.
Otherwise it is exactly the same (same events, same body loading,
etc).