dropdown: content elem and body init changes
This updates how body content is initialized.  The style class attr
is still always applied to the body elem.  However, the size attrs
are applied now to a _content_ elem.  This content elem is optional;
size fallback-applies to the body elem if there is no explicit
content elem.

This allows you some level of control on what markup gets the size
attrs.  By default, the body does.  However this allows some custom
body sub-markup to also be able to get it.

Use this to add header/footer sections to dropdown content and keep
those section from being affected by sizing or overflow scrolling.