indicator: component for indicating something is happening
This adds a new component, indicator, for indicating something is
happening in the UI.  It has a `start` and `stop` API to turn it
on and off as needed.

This uses spin.js to create a spinner to use as the UI indicator.
This follows the pattern on our UI components.  You bind it to an
element.  The spin js spinner opts are set based on the element
you are binding to.  Events are emitted as stuff happens along the
way.

A big inteded use for this is on our forms, indicating when a form
is being ajax-submitted.  Therefore, the forms auto trigger their
submit elems.  This is totally optional - just don't setup the
submit as an indicator and it won't be triggered.