modals: drag n drop to reposition
This adds drag n drop support to modals.  Define an element in the
modal popup with `data-romo-modal-drag="true"` and that elem becomes
a "handle" for dragging the modal.

A modal's position is fixed, so you can drag anywhere in the viewport
and the modal will "stick" there instead of the default "center"
position.  Window resizes trigger a reset and position the modal
back in the center of the viewport.

This is mainly b/c modals overlay content and may sometimes block
content that you need to see to interact with the modal.  This allows
users to move the modal out of the way if needed.