dropdown, modal: timeout popup open/close to ensure event order
Events are bound on popup open/close.  When one popup closes, it
needs to unbind its events before another popup binds its events.

However, when clicking off the popup triggers a different elem's
popup to open, then open event was process before the click off
(because the invoke event is handled before the event can bubble up
to the body to trigger the click off).  This caused events to not
fire in the correct order and messed up event binding in this
specific scenario.

This adds timeouts to opening/closing popups when triggered from
events.  This timeout allows the event to bubble up and trigger
the click off (and the unbind) before opening and binding.

This is a very edge case scenario but when event binding messes up
it causes the UI to break.  This keeps the UI from breaking in this
scenario.