modal: position top at a perc height instead of centered.
This switches to positioning the top of the modal at a percentage
of the viewport height instead of centering it in the viewport
height.  This ended up feeling more natural than always centering.

If the height of the modal needs to be less than the percent height,
it falls back to centering - this should only be needed in the case
of modals with "tall" content.