select: first pass at select component js/css; select UI
This is the first pass/start of the new select component.  This
sets up the basic component structure.  The component replaces the
normal select UI with a button-dropdown.  This covers the js/css
for the the button-dropdown, the basic button UI event handling
and basic button UI css.

There is more to do before the component is ready: caret UI, popup
handling and UI, general input handling and interactions, etc...