select: add icon for caret
This adds settings to specify an icon to use as a caret.  This also
adds the css/js to handle positioning styling the caret.

This uses a graduated scale percentage width scheme to position
the caret.  This has the advantage of being able to use vertical
align middle styling to keep things lined up nice.  The graduated
scale makes positioning work at extremely small and large sizes.