dropdown: update to work with new select component
The new select component has some specific features it needs the
dropdown to provide.  This updates dropdown to implement those
features and updates select to use them.

The first feature needed is the dropdown popup needs to match the
width of the select UI.  By default, the popup is sized based on
its content.  This adds a special setting, `elem`, which directs
the dropdown to match the popup width to the element width.

The second feature is how overflow-x is handled.  By default,
dropdown scrolls overflow x.  This adds an option to specify both
overflow x and y.  The select then forces its overflow x to hidden.

Finally this includes a few style updates to support these features.

* the dropdown popup needs to