select: trigger the change event when item is selected
This adds triggering the 'change' event on the select when an item
is selected.  It is given the new value as an arg.

This also updates the 'select:itemSelected' event to receive both
the new value and the prev value as args.  This will allow users
to do conditional logic on values changing from one thing to
another.