select: only trigger change events if value actually *changes*
Previously an event would be triggered whether the value actually
changed or not.  This switches to calculating the prev and new vals
beforehand and then only running the logic if the values are
different.  This is both more efficient and appropriate.

This also changes the default 'change' event to trigger with no
args and updates the custom change event name to 'select:change'.
This is more standard and appropriate for these events.