select: force update to the selected option on page show
This is to help fix weird browser handling on page show (particularly
when using the back button).  This ensures that the option marked
as selected is always show as selected in the component UI.

Without this, browser on the back button renders selects as they
were just before the page change, not as they were originally
rendered.  This ensures selects show up as originally rendered
on back button.