dropdown: bind to focused escape to manually close the dropdown
The dropdown auto closes on escape, however, it uses the same logic
as other popup components like the modal.  This works fine unless
you nest dropdowns in modals.  This means that when you hit esc on
a dropdown in a modal, it will close both the dropdown and the modal.
This is not ideal.

This updates the dropdown logic to also bind to esc on the elem
and popup elems themselves and if the popup is open, manually close
it and stop propagating the event.

This means hitting esc on a focused dropdown won't close other
components as well as itself.  However, it keeps the behavior of
hitting escape when not focused closing everything.