modal/dropdown: rework body click handling to allow proper interaction
This came up while nesting a dropdown in a modal.  Because the
dropdown popup is not contained in the modal popup, clicking off
the popups was not closing them as expected.

So, because modals are designed to contain dropdowns but dropdowns
are not designed to contain modals, we can change the click on body
handling to fix this.

First we add checks to both body handlers to ensure they only close
the popup if it clicked element is not a child of the popup.  Then
we keep the popup click event that stops propagation on the dropdown
but we _remove_ it in the case of the modal.  We need to the click
event to propagate to the window body for modal clicks b/c that is
what will close any nested dropdowns so we can't stop propagation
on it.  However, we need to keep stopping propagation on the dropdown
b/c if we didn't, when clicking on popup UI nested in a modal, the
event would propagate to the window and close the modal.

Anyway, this gets the nested popup close handling updated to work
as expected.