modal/dropdown/tooltip: updates for proper nesting/interactions
This updates the modal to trigger body events when certain actions
take place: when mousedown to start dragging and when closing the
popup.  These events need to be triggered specially b/c they help
notify other elements of a modal popup change.

The mousedown event needs to be triggered explicitly b/c the handler
stops event propagation.  Without this, the event could never be
handled by other components.

The popupclose event needs to be triggered as a generic way to
signal other elems that the popup is closing.

This is all so that any dropdown/tooltip popups generated by elems
nested in modals, will behave properly when a modal is dragged or
closed.

* the dropdown sets it's position to fixed to match the modal popup
  (this is to keep the dropdown positioned in sync with the modal)
* the dropdown closes it's popup when a modal mousedown happens (this
  is needed b/c the mouse down event triggers before the body click
  event - without the dropdown popup remains shown while modal dragging)
* the tooltip hides itself on mousedown or popupclose if nested in
  a modal.  this is needed b/c tooltips can remain displayed if their
  delay is high enough.

This is all just tweaks to get these components to honor each other
and interact as expected.