inputs: basic css
This adds basic css for styling romo inputs.  All inputs require the
'romo' class to turn on styles so it is safe to use alongside default
input styling.

This is prep for adding the datepicker.  The datepicker will be
used on text inputs so we need to have them styled as expected so
things work nicely.