dropdown: add 'disable toggle' feature; use in datepicker
This adds a markup setting to have dropdowns disable their toggle
feature.  This means the they will open on toggle click but not
close on toggle click.  This setting is now forced on datepicker
components.

This is useful for any dropdown that uses a text input as a toggle.
These are designed to be clicked on and text double-clicked, etc.
You don't want you dropdown show/hiding constantly as the user
clicks around in the input.

Users can still dismiss the dropdown using the escape key or by
clicking on the body like normal.