dropdown/tooltip: better z-index parsing and handling
So some browsers suck and can't handle looking up inherited z-index
values using the `.css()` method.  This adds a base helper for
parsing an elements z-index value that manually walks the parents
and grabs the first numeric z-index value.

This is then used when setting the z-index on dropdowns and tooltips.
These components are designed to nest in modals so they need to set
their index value relative to any parent index value.

I found this by loading a select dropdown in a modal in Chrome, btw.
Firefox worked great and "figured it out".