datepicker: setup refreshing UI on dropdown open event
This is setup for rendering the calendar based on the current date
value for the component.  This date is set either on init or when
a new value is selected.  When the popup is opened, the calendar
will be refreshed using the current date or today's date if there
is no date.

This is all interaction and setup for rendering the calendar dyanmically
based on the date value.