datepicker: update build calendar logic to accrue html content
This updates the build calendar logic to accrue its html markup
and append it to the body.  This allows us to iterating over
dates and accrueing markup along the way.

In addition, this adds a helper function for getting a utc date
based on given year, month, and day values.

All of this will be used to generate calendar markup for a given
date's month.