datepicker: allow passing date to refresh; store refresh date on refresh
This updates the refresh UI logic a bit to prep for cycling prev/next
months in the calendar UI.  To implment this feature, the refresh
logic not only needs to take an arbitrary value to refresh to (falling
back to its defaults) but to also store the date it was last refreshed
to.

Using these features we can implement the prev/next logic.