datepicker: more proper selected style behavior
Before this would make whatever data was used to refresh the UI
as selected.  This happened to work on the initial refresh.  But
once we implement the prev/next logic, this logic will no longer
work.

The proper solution is to only mark selected the value that
the component says is selected, the `this.date` (if it exists).
This has the side benefit of not selecting any value if launching
the popup with an empty input value.

This is a cleanup that I found while implementing the prev/next
logic.