datepicker: force a min width on the calendar UI
This updates how the calendar dropdown width is handled.  Normally,
the dropdown matches the width of the elem.  However, if the elem's
width falls below a threshold, the calendar width stays the same
min width and the elem is just narrower.

This keeps the UI from breaking in the case of using a very narrow
datepicker.