select: switch to abs position caret; more like how datepicker indicator
This switches to using abs positioning to position the select caret
in the select UI.  This is more consistent with how datepicker
positions its indicator.

The comes with some CSS overhead, but the main advantage is that
the select UI looks more consistent at different widths.  Before,
with using percentage-width positioning, we had to have this
graduated scale for specifying the widths.  It was hard-coded and
manual and just didn't look that great.

In addition, this positions the select caret and the datepicker
indicator with paddings that match the inputs (6px).  This padding
for these has been adjusted as well (22px).