dropdown: don't return false on popup body clicks
This prevents "default behavior" like clicks on anchors in the
popup body from happening.  The stop propagation is sufficient
to prevent the event from bubbling to the body and closing the
dropdown.

This had been added in 0e0db0df53ebdc224adb379ac3d70a9b3414de31
in an attempt to make dropdowns nested in modals behave properly.
The return was in fact not necessary to accomplish what that
commit set out to do - the other stuff in that commit is sufficient.