datepicker: update focus behavior
We can't use `select` to trigger opening the popup as it doesn't
work consistently in all browsers.  So this changes the behavior
to only opening the popup on focus if the elem val is empty.

This ensures that the popup doesn't re-open after you select a
value and ensures that it does open when focused on all browsers.