gh-pages: add a build script for building gh-pages static files
This adds a script to build all the static files needed to host
the gh-pages.  Running this outputs all content to the tmp dir.
Checkout the gh-pages branch and update from the tmp dir.

This uses scmd to start up the deas server, then uses wget to
recursively download the static files to the tmp dir.  Once everything
is downloaded, it stops the server.

This also updates the server to create links and such with the
appropriate base url for hosting on gh-pages.