gh-pages: test pages setup
This setups up rendering test pages.  In addition, it also adds
a Kramdown test page.  Kramdown will be used to render markdown
in our pages.

Since we will want to display a lot of example code in the site,
this also adds in a coderay stylesheet to render syntax highlighting.
The stylesheet tries to mimic the Github syntax highlighting as
much as possible.