css: setup typography CSS
This renames the globals css to `base.css` and adds some basic
body font and paragraph settings.  This is setup for additional
typography CSS settings.

This tweaks the gh-pages custom css to use the coderay defaults
and then override them for this site in `gh-pages.css`  This is
to force font sizes regardless of what Romo might use as defaults.

This also updates the gh-pages site to have a typography page to
document these settings.