css: use Scss variables in base CSS
This sets up using the Scss format for the base css.  This allows
us to define and import some variables when compiling.

This also switches to using variables to define some base typography
settings: font family, size and line height.

To accomplish this, we need to update to the latest dassets-sass
in the gh-pages site.  This version properly sets the load path for
`@import` calls.

This also updates to the latest coderay.css.