css: have bordered grid table cells hide overflow
Since cells in a row don't have uniform height, if one cell overflows
it breaks the borders for its neighboring cells that don't overflow.

To handle this we just bail and don't allow any overflow.  If bordered
overflow is needed, it is better to use a normal table or don't use
a bordered grid table.