add an optional `romo/dassets` require to do standard Dassets build
This adds an optional require to do the Dassets configuration for
the standard Romo build.  This is for use in Dassets build
configurations.

This is a major prototype and won't live here forever.  It should
eventually be moved to a romo-dassets gem or something.

This switches the gh-pages site to use it and validate it.