css: tweak inline lists to only pad one side between items
This avoids a "double pad" scenario that is only obvious when not
bordering the items.

As a side effect of this change, I removed the border style classes
as they aren't useful when only padding one side.  If this is needed,
you should have the item content pad both sides and border as needed.