css: tweak tabs to properly display border when stacked
This fixes some broken CSS I missed when developing stacked tabs.
You want the border to show when you stack tabs, but not with
regular tabs.

This broke b/c I adjusted reg tabs to show border but just match
the bg color.  This is so their height would match buttons.  This
is what broke the stacked tabs border.

In addition, this also tweaks the anchor handling to have it
not switch black when you click on an anchor in the stacked tabs.