css: tweak default striped and alt bg colors and usage
Using the straight "alt" color as the "striped" color on grid tables
and tables wasn't good enough.  It locks you into something you may
not want.  Typically, you want the striped color to me close to the
base color so the striping isn't so harsh.  But this may not always
be true for the alt color.  There, having harsher contrast with the
base color is desired.

This also cleans up the table docs a bit to make them more consistent
with the grid table docs.