form-js: check the submit elem for disabled class, not the form
This switches to checking the form's submit elem for the disabled
class when it's clicked, instead of checking the form. Otherwise,
you would need to add a disabled class to both the submit elem (to
get the styles) and the form (to disable submitting). Add a disable
class only to the button is more consistent with disabling buttons
for modals and other components.