form js: don't bother with building arrays of input values
Only build arrays when the input is marked with
`data-romo-form-list-values="true"`.  This allows you to build arrays
like it would normally.  With this change only the last value of
same-named-inputs will be sent.  This allows you to do the
hidden-input-true-false-checkbox trick.  If you need to send multiple
checkbox values for the same input name, just use list-values-true.