dropdown/modal js: calculate pad bottom for max height detect
This updates the modal and dropdown popup handling to calculate
how much bottom space is needed in the max height detect calculations.

Before this was just a default `10` pixels of buffer.  But if your
content is in a nested elem with other elems below it, this breaks.
You need to calculate the difference between the content bottom
and the body bottom and factor that into the calculations.  If the
body is the content, then it the bottom space will be zero and
this behaves as before.