select: make the caret feel more integral to the component
Previously, the hover styles would go away if you hovered the caret.
Also, you couldn't click on the caret and open the dropdown.  Also
the caret wouldn't get the default typography styles the rest of
the component did.

This fixes all that and make the caret feel like it is a part of
the component.