css: remove border sizes 3-6 - switch to new standard 0-2 sizing scheme
The 1-6 size options seemed arbitrary and a little excessive.  This
scheme limits the choices and has them variable driven so they can
be customized.

This will be the new sizing scheme for borders, border radius, pad,
push, text, buttons, etc - anything that is sized.

The design behind this scheme is the `1` is the default size. `0`
is "small" and `2` is "large".  Sometimes small and large will be
aliased, sometimes not.

Anyway, this gets borders switched to this scheme.