js: add basic sortable component
This adds a basic sortable JS component. This follows the
conventions of the other JS components and uses HTML5 drag-and-drop
logic. This allows users to sort objects in a list and triggers
an event as objects are sorted.

This also commonizes styles for grab cursor, grabbing cursor and
user select. This is to keep the styles maintainable and keep from
duplicating styles in the CSS.