css: switch text, line-height, img, button, label, tab size to 0-3 scheme
Originally they were sized on a 1-4 scheme with 1 being biggest (intended
for use on `h1`s).  This was short-sighted in hindsight.  Since we have
switched all other sizing over to the 0-2 scheme, this should match that
in a similar fashion.

Here 0 is smallest and 3 is biggest.  The reason we go to 3 on the size
numbers is b/c you don't want to use size 0 in the headings and you need
3 sizes to apply to the headings.  So, we add size 3 to accomodate this.

This gets the sizing more closely in line with the other sizing scheme.
This also tweaks the line-heights and image sizes to make them more
consistent and work better with corresponding text sizes.  These are
just some cleanups that need to be made now that Romo has been in use
for a bit.