select js: better item click handling
The recent blur logic I added jacks up selecting items if you "hold
and click" on them.  The blur fires and closes the popup before the
click event registers.

This switches to putting the blur on a timeout (to ensure any mouse
down happens before it).  Then I add a popup mouse down/up to set
state on whether you are mousing down on the popup.  Then the blur
logic checks and doesn't close the popup if you are mousing down
on it.  Finally the click comes through and clears any timeout and
stops propagation to be explicit about its handling.

All this makes for consistant item click behavior.