gh-pages: docs layout with links sidebar; CSS handler helper
This adds a new layout for the doc pages themselves.  This layout
has a links sidebar and provides methods for adding links to the
sidebar programmatically.

This also adds a css handler helper mixin.  This helper defines
the docs links for the css pages.  All the css handlers are now
switched to just including this helper.

Finally, this adds a bit of js to, on page load, calc the sidebar
width and manually set that and make it position fixed.  This needed
to be done with js b/c setting position fixed without a width causes
the sidebar not to fill up its width.  This had to be done on page
load so the width will be calculated correctly.

This is part of sprucing up to docs pages for an initial publish
in addition to making it easier to navigate the css docs.