gh-pages: css show page; links on main index and nav bar
This adds a show page for the CSS section of the docs and then
links to it from the main index page.  This also updates/corrects
the links in the nav bar.

This is part of getting the site ready to publish.  These pages
are bare-bones for now but add links so that the site can be built.
In future efforts I'll spruce them up a bit.