button css: remove gradients from buttons due to poor browser performance
Chrome has issues rendering large numbers of gradients so I've
decided to remove this from buttons.  Plus this feels kinda
nice having a flat-but-not-too-flat design feel to go with the
default-square border radiuses on elements.  Anyway, this is mostly
about performance so you can use a ton of buttons on a page.