datepicker js: blur closes popup; better click handling
This updates the datepicker component to close the popup when it is
blurred.  However, this logic jacks up any click logic if you "hold
and click".  The blur fires and closes the popup before the click
event registers.

To account for this, the blur logic is in a timeout (to ensure any
mouse down happens before it).  Then the popup mouse down/up sets
state on whether you are mousing down on the popup.  The blur logic
checks and doesn't close the popup if you are mousing down on it.
Finally any clicks clear any blur timeout and stops propagation
to clean things up.

All this makes for consistant item click behavior while being able
to close the popup on blur.  This follows the pattern and logic on
the select component.