dropdownForm js: initial component
This adds a component for invoking a dropdown that contains a form.
This follows the pattern of the similar inlineForm and modalForm
components.

This also updates the dropdown component to be aware of "close
elements" - elements that when clicked close the dropdown.  This
follows a similar pattern used by modals.

All this is to allow for easily nesting forms in dropdowns, much
like you do in modals, and having them interact nicely.