button css: force height on sized buttons
This updates the button css to add a mixin for calculating button
height and setting that in addition to line-height.  This fixes
a bug where vertically aligning child elements in a button causes
the height to increase by ~3px.  We now lock the appropriate height
based on the line height and it fixes the problem.