gh-pages: update to latest Deas; move urls to web layout handler
This moves all url definition into the web layout.  This gets
logic out of the web template.  This also updates to the latest
Deas.

This is needed b/c the latest Deas has deprecated the `router` and
`url_for` methods that handled prepending the base url to raw url
paths.  The advantage here is there is less logic in the template
and the urls can be tested.