forms js: specify indicator elems like you do a submit elem
This updates the form js components to pass in any indicator elements
they know about.  The form now triggers those indicator elements
(instead of the submit elem).  Any indicators in the form will be
used if none are given.  This approach mirrors how the submit elem
is determined.

This allows you to use other elements besides the submit elem for
indicating form submission.