Add `romo-img-inline` classes for lining up text and images
This adds `romo-img-inline` classes for lining up text and images
in specific scenarios. When the text is smaller than the image
and the elements are not inline (typically because of using the
grid system) the text will not be vertically aligned with the
image. This allows manually setting the text's line height to the
images height so they will vertically align.