select js/css: switch to using romo-btn UI
This switches the generated select component UI to use `romo-btn`
styles.  This keeps the UI consistent with the button UI (which is
the intent) and also removes the need for a whole bunch of styles.

This also tweaks the remaining styles to use appropriate variables
and remove styles that are no longer needed.  The select component
was written in a vaccuum before the CSS so it never took advantage
of it before.