Change button and label inline classes to not set font size/weight
This changes the button and label inline classes to not set font
size or weight and only set line height. This is to make the
classes more flexible and useful when mixing sizes with buttons,
labels and text. This makes them consistent with the image inline
and input inline classes.