select css: break out select mixins; apply to component and form css
This takes the stylings for the select component, breaks them off
into mixins, and then reuses them on the main form css.  This means
that selects in the main for and the component will look similar.

This also includes some cleanups to the input variables and updates
the docs to show the styles on a multi grouped select.