js: select dropdown component
This breaks out the dropdown handling/building/behavior from the
select component as its own component.  The goal here is you can
render dropdowns as if they were a select's dropdown but without
having to have a select element.

Right now, this just seperates the behavior and logic.  The select
now composes a select dropdown.  There is (or should be) no behavior
changes to the select component itself.  You still use it as before.