select dropdown: updates to work better standalone from a select
This updates the select dropdown to now optionally take an options
parent element.  If none is given is will look for an element with
a certain class and use that as the options parent element.

This way a select can manually pass in its options parent element
as before while standalone you just define the options under a
parent with a certain class.

Also, this moves the logic for hitting "Enter" to select an option
to a more reliable event handler (the body).  This way is you lose
focus on the elem/popup, hitting enter will still select the
hightlighted item.

Finally, this moves the elem focus behavior out of the select-dropdown
and into the select component.  If we are in a select it is safe
to assume we want to focus the input after an item is selected.
However, if we are in a select-dropdown we shouldn't assume that for
the user.  They can do this themselves if they see fit to.